requirements.txt
appdirs==1.4.3
colorama==0.3.9
decorator==4.0.11
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-libnmap==0.7.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.9
traitlets==4.3.2
wcwidth==0.1.7
